port `agent.js` tests to `node:test` runnner
